this patch fixes a bug in the queue agent handling code where there is a possible buffer overrrun.  when agents are notified of a waiting message, they are given the size of the envelope.  this conversation was not properly protected against an overrun.
the code should be slightly slower, and use a more memory, and it might not be fully patched.

todo:  there are three places in code where the envelope is read.  the two existing ones should have already been fairly protected, but they don't use connio to its 
potential.  fixing that will allow for the removal of PullLine() and the exclusive use of PullLine2()